requirements.txt
# Core dependencies
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing and technical analysis
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.28  # Technical Analysis Library

# Optional: for enhanced data handling
python-dateutil>=2.8.2
pytz>=2023.3

# Logging and monitoring
colorlog>=6.7.0

